Add a production server so that the app can be deployed
* Add a production version of the Webpack config file
* Add a production server that does not contain webpack-hot-middleware

The biggest challenge here is that between development and production
mode, the path for the bundle file that Webpack produces is different
(in prod, it contains a hash, whereas in dev, it does not). So in our
main HTML file, when we are bringing in that bundle file, we need the
path to reflect the name that Webpack gives to it. If we weren't using a
server, we could have used HtmlWebpackPlugin, which will automatically
create an HTML file for us with the right path for the bundle file. But
since we're not, we have to do this ourselves.

Fortunately, Webpack has the ability to produce a manifest file which
contains all of the file names that it has given to all of the files
it's compiled. This file will exist in both dev and prod modes, so in
the server code, we can read the manifest and pass in the file path to
the index view that Express renders.