Use CSS Modules to style React components
There are obviously a plethora of different options for styling React
components using JavaScript[1]. Some of the most popular ones from this
list include Aphrodite, Emotion, Glamorous, Stylable, and
styled-components (I'm sure I'm missing a few). All of them were
more or less inspired by this talk[2].

However, I don't like the CSS-in-JS trend. Don't get me wrong: the idea
behind it isn't necessarily bad -- it's a good idea to "componentize"
CSS. But they all seem to espouse shoehorning CSS either into a string
or a mixture of an object and strings. It seems wrong.

See, I've been using BEM for a while now -- I think it's an elegant and
yet incredibly simple solution for corralling your CSS and curtailing it
around the component pattern. But there are obviously limitations. For
one, it may be *too* simple -- you and your teammates have to be
diligent enough to map your CSS files one-to-one with your components.
It also becomes cumbersome when combined with React, because you have to
end up repeating the name of your component over and over again. Plus,
at the end of the day, the CSS for your component is global, when it
really should be local to your component.

Given this, I feel like CSS Modules solve these problems in a way that
allows me to keep writing CSS just like I've always done, only now it's
tailored for components. This article[3] by Glen Maddern, the creator of
CSS Modules, is a great read and really clarified all of the things that
the "anti-CSS-in-JS" philosophy brings to the table. I've added the
react-css-modules package in this commit, which is a big help.

[1]: http://michelebertoli.github.io/css-in-js/
[2]: https://speakerdeck.com/vjeux/react-css-in-js
[3]: https://glenmaddern.com/articles/css-modules