Update server to lazily read Webpack manifest
When running the dev server for the first time, if we try to read the
manifest.json file that Webpack generates, it won't work, as the
`assets` folder will not exist until Webpack has a chance to run. To fix
this, defer reading manifest.json until the index page is served, and
cache the result so that we don't end up reading the file over and over.