Rename Lattice to Honeycomb
It's true that we're drawing a lattice, but it's going to prove to be
helpful to make a Lattice model as well, and rather than sharing the
name it's less confusing to make up a new name. When we draw the lattice
we end up making a honeycomb structure, so that's why we chose this
name.