Move app/services to app/models
I've used "services" to hold non-component related functions and classes
in the past -- I stole it from Angular -- but it's not a great name. All
of the code here is used to model the underlying data that powers the
visuals, so "models" is an appropriate name. If I need to make another
folder later to hold other things that's totally fine with me but this
works for now.