Prevent lattice from getting cut off at the top
In order to render the lattice, we have to calculate the pixel position
for each cell in the lattice based on its location in the underlying
data structure. Some cells will have a negative top position and will
therefore get chopped off. If this happens, adjust the top position of
the entire lattice to keep all of the cells in view. (If the lattice
exceeds the bottom edge of the viewport, then we can have the user
scroll downward to see the rest of it, but if it exceeds the top edge of
the viewport, the user is not able to scroll up.)