Make lattice more useful
Prior to this commit we would normalize each ratio so that every note
falls within one octave. However, you can't play this version of the
lattice very well, because not all major and minor triads are
represented as the same inversion.

Instead of normalizing each ratio, multiply and divide notes by either
5/4 or 3/2.