Ensure that the AudioContext is started
Some browsers require that the user provide some kind of input (like a
mouse click or tap) before playing any sound. Every so often, I've
noticed that sound doesn't work when you load the page, and whenever
this happens we're getting a warning from the browser in the console.

The `startaudiocontext` package ensures this by binding to the `mouseup`
or `touchend` event of the element you provide and then starting the
AudioContext thereafter. Hopefully it will solve this problem.